requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_190_select=0.0.1=gpu
absl-py=0.3.0=py_0
alabaster=0.7.12=py27_0
asn1crypto=0.24.0=py27_0
astor=0.7.1=py_0
astroid=1.6.5=py27_0
attrs=19.1.0=py27_1
babel=2.7.0=py_0
backports=1.0=py27_1
backports-functools-lru-cache=1.5=pypi_0
backports.functools_lru_cache=1.5=py_2
backports.shutil_get_terminal_size=1.0.0=py27_2
backports.weakref=1.0.post1=py27_0
backports_abc=0.5=py27_0
blas=1.0=mkl
bleach=1.5.0=py27_0
bzip2=1.0.6=h470a237_2
ca-certificates=2019.5.15=0
cairo=1.14.12=he56eebe_3
certifi=2019.3.9=py27_0
cffi=1.11.5=py27h9745a5d_0
chardet=3.0.4=py27hfa10054_1
click=7.0=pypi_0
cloudpickle=0.6.0=pypi_0
configparser=3.7.4=py27_0
cryptography=2.4.1=py27h1ba5d50_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.1.2=cuda9.0_0
cupti=9.0.176=0
cycler=0.10.0=pypi_0
dash=0.21.1=pypi_0
dash-core-components=0.23.0=pypi_0
dash-html-components=0.11.0=pypi_0
dash-renderer=0.13.0=pypi_0
dask=0.19.4=pypi_0
dbus=1.13.0=h3a4f0e9_0
decorator=4.3.0=py27_0
defusedxml=0.6.0=py_0
docutils=0.14=py27hae222c1_0
entrypoints=0.3=py27_0
enum34=1.1.6=py27h99a27e9_1
expat=2.2.5=hfc679d8_1
ffmpeg=4.0.2=ha6a6e2b_0
flask=1.0.3=pypi_0
flask-compress=1.4.0=pypi_0
fontconfig=2.13.0=hd36ec8e_5
freetype=2.8.1=hfa320df_1
fuel=0.2.0=pypi_0
funcsigs=1.0.2=py_2
functools32=3.2.3.post2=pypi_0
futures=3.2.0=py27h7b459c0_0
gast=0.2.0=py_0
gettext=0.19.8.1=0
giflib=5.1.4=h470a237_0
glib=2.55.0=h464dc38_2
gmp=6.1.2=hfc679d8_0
gnutls=3.5.19=h2a4e5f8_1
graphite2=1.3.11=hfc679d8_0
grpcio=1.12.1=py27hdbcaa40_0
gst-plugins-base=1.12.5=hde13a9d_0
gstreamer=1.12.5=h61a6719_0
h5py=2.8.0=pypi_0
harfbuzz=1.8.5=h2bb21d5_0
hdf5=1.10.2=hc401514_1
html5lib=0.9999999=py27_0
icu=58.2=hfc679d8_0
idna=2.6=py27h5722d68_1
imageio=2.4.1=pypi_0
imagesize=1.1.0=py27_0
imbalanced-learn=0.4.3=py_0
intel-openmp=2018.0.3=0
ipaddress=1.0.22=py27_0
ipykernel=4.10.0=py27_0
ipython=5.8.0=py27_0
ipython_genutils=0.2.0=py27h89fb69b_0
isort=4.3.20=py27_0
itsdangerous=1.1.0=pypi_0
jasper=1.900.1=4
jedi=0.13.3=py27_0
jinja2=2.10.1=py27_0
jpeg=9c=h470a237_0
jsonschema=3.0.1=pypi_0
jupyter_client=5.2.4=py27_0
jupyter_core=4.4.0=py27_0
kiwisolver=1.0.1=pypi_0
lazy-object-proxy=1.4.1=py27h7b6447c_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran=3.0.0=1
libgfortran-ng=7.2.0=hdf63c60_3
libiconv=1.15=h470a237_2
libopenblas=0.2.20=h9ac9557_7
libopencv=3.4.1=h1a3b859_1
libpng=1.6.35=ha92aebf_0
libprotobuf=3.5.2=h6f1eeef_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he6b73bb_1
libuuid=2.32.1=h470a237_0
libwebp=0.5.2=7
libxcb=1.13=h470a237_2
libxml2=2.9.8=h422b904_3
lxml=4.2.5=pypi_0
markdown=2.6.11=py_0
markupsafe=1.1.1=py27h7b6447c_0
matplotlib=2.2.3=pypi_0
mccabe=0.6.1=py27_1
mistune=0.8.4=py27h7b6447c_0
mkl=2018.0.3=1
mkl_fft=1.0.5=py27_0
mkl_random=1.0.1=py27h4414c95_1
mock=2.0.0=py27_0
nbconvert=5.5.0=py_0
nbformat=4.4.0=py27_0
nccl=1.3.5=cuda9.0_0
ncurses=6.1=hf484d3e_0
nettle=3.3=0
networkx=2.2=pypi_0
ninja=1.8.2=py27h6bb024c_1
numexpr=2.6.6=pypi_0
numpy=1.15.0=py27h1b885b7_0
numpy-base=1.15.0=py27h3dfced4_0
numpydoc=0.9.1=py_0
olefile=0.46=py27_0
openblas=0.2.20=8
opencv=3.4.1=py27h6fd60c2_2
openh264=1.7.0=0
openssl=1.1.1=h7b6447c_0
packaging=19.0=py27_0
pandas=0.24.2=py27he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py27_1
parso=0.4.0=py_0
pathlib2=2.3.3=py27_0
pbr=4.2.0=py_0
pcre=8.41=h470a237_2
pexpect=4.6.0=py27_0
picklable-itertools=0.1.1=pypi_0
pickleshare=0.7.5=py27_0
pillow=5.1.0=py27h3deb7b8_0
pip=10.0.1=py27_0
pixman=0.34.0=2
plotly=3.10.0=pypi_0
progressbar2=3.38.0=pypi_0
prompt_toolkit=1.0.15=py27h1b593e1_0
protobuf=3.5.2=py27hf484d3e_1
psutil=5.6.2=py27h7b6447c_0
pthread-stubs=0.4=h470a237_1
ptyprocess=0.6.0=py27_0
py-opencv=3.4.1=py27h0676e08_1
pycodestyle=2.5.0=py27_0
pycosat=0.6.3=py27ha4109ae_0
pycparser=2.18=py27hefa08c5_1
pyflakes=2.1.1=py27_0
pygments=2.3.1=py27_0
pylint=1.9.2=py27_0
pyopenssl=18.0.0=py27_0
pyparsing=2.2.2=pypi_0
pyqt=4.11.4=py27_4
pyrsistent=0.15.2=pypi_0
pysocks=1.6.8=py27_0
python=2.7.15=h9bab390_4
python-dateutil=2.7.3=pypi_0
python-utils=2.3.0=pypi_0
pytorch=0.4.0=py27hdf912b8_0
pytorch-cpu=1.0.1=py27hf484d3e_0
pytz=2018.5=pypi_0
pywavelets=1.0.1=pypi_0
pyyaml=3.13=pypi_0
pyzmq=17.1.0=pypi_0
qt=4.8.7=2
qtawesome=0.5.7=py27_1
qtconsole=4.5.1=py_0
qtpy=1.7.1=py_0
readline=7.0=ha6073c6_4
requests=2.18.4=py27hc5b0589_1
retrying=1.3.3=pypi_0
rope=0.14.0=py_0
ruamel_yaml=0.15.37=py27h14c3975_2
scandir=1.9.0=py27h14c3975_0
scikit-image=0.14.1=pypi_0
scikit-learn=0.19.2=pypi_0
scipy=1.1.0=pypi_0
setuptools=39.2.0=py27_0
simplegeneric=0.8.1=py27_2
singledispatch=3.4.0.3=py27_0
sip=4.18=py27_0
six=1.11.0=py27h5f960f1_1
snowballstemmer=1.2.1=py27_0
sphinx=1.8.5=py27_0
sphinxcontrib=1.0=py27_1
sphinxcontrib-websupport=1.1.2=py_0
spyder=3.2.8=py27_0
spyder-kernels=0.4.4=py27_0
sqlite=3.25.3=h7b6447c_0
subprocess32=3.5.3=pypi_0
tables=3.4.4=pypi_0
tensorboard=1.9.0=py27hf484d3e_0
tensorflow=1.9.0=gpu_py27hd3a791e_1
tensorflow-base=1.9.0=gpu_py27h6ecc378_0
tensorflow-gpu=1.9.0=hf154084_0
tensorlayer=1.10.1=pypi_0
termcolor=1.1.0=py_2
testpath=0.4.2=py27_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=pypi_0
torchvision=0.2.1=py27_1
tornado=5.1.1=py27h7b6447c_0
tqdm=4.25.0=pypi_0
traitlets=4.3.2=py27_0
typing=3.6.6=py27_0
urllib3=1.22=py27ha55213b_0
wcwidth=0.1.7=py27h9e3e1ab_0
webencodings=0.5=py27_0
werkzeug=0.14.1=py_0
wheel=0.31.1=py27_0
wrapt=1.10.11=pypi_0
wurlitzer=1.0.2=py27_0
x264=1!152.20180717=h470a237_0
xorg-kbproto=1.0.7=h470a237_2
xorg-libice=1.0.9=h470a237_4
xorg-libsm=1.2.2=h8c8a85c_6
xorg-libx11=1.6.5=h470a237_2
xorg-libxau=1.0.8=h470a237_6
xorg-libxdmcp=1.1.2=h470a237_7
xorg-libxext=1.3.3=h470a237_4
xorg-libxrender=0.9.10=h470a237_2
xorg-renderproto=0.11.1=h470a237_2
xorg-xextproto=7.3.0=h470a237_2
xorg-xproto=7.0.31=h470a237_7
xz=5.2.4=h470a237_1
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=ha838bed_2
